refactor: Icon component
