+++ src/components/Icon/helpers/getIconClasses.ts
@@ -3,19 +3,15 @@ import { IconClassType } from '@/styles/constants';
 import { IconClassSet } from '../types';
 
 export const getIconClasses = (type: IconClassType): IconClassSet => {
+  const baseCircleClass =
+    'relative flex items-center justify-center rounded-full';
+  const baseIconClass = 'absolute';
+
   switch (type) {
     case IconClassType.SERVICES:
       return {
-        circleClass: `
-          w-[68px] h-[68px]
-          tablet:w-[80px] tablet:h-[80px]
-          desktop:w-[108px] desktop:h-[108px]
-        `,
-        iconClass: `
-          w-[64px] h-[64px]
-          tablet:w-[72px] tablet:h-[72px]
-          desktop:w-[96px] desktop:h-[96px]
-        `,
+        circleClass: `${baseCircleClass} w-[68px] h-[68px] tablet:w-[80px] tablet:h-[80px] desktop:w-[108px] desktop:h-[108px]`,
+        iconClass: `${baseIconClass} w-[64px] h-[64px] tablet:w-[72px] tablet:h-[72px] desktop:w-[96px] desktop:h-[96px]`,
         offsets: {
           mobile: 24,
           tablet: 30,
@@ -25,16 +21,8 @@ export const getIconClasses = (type: IconClassType): IconClassSet => {
 
     case IconClassType.EXPERTISE:
       return {
-        circleClass: `
-          w-[68px] h-[68px]
-          tablet:w-[80px] tablet:h-[80px]
-          desktop:w-[64px] desktop:h-[64px]
-        `,
-        iconClass: `
-          w-[64px] h-[64px]
-          tablet:w-[72px] tablet:h-[72px]
-          desktop:w-[64px] desktop:h-[64px]
-        `,
+        circleClass: `${baseCircleClass} w-[68px] h-[68px] tablet:w-[80px] tablet:h-[80px] desktop:w-[64px] desktop:h-[64px]`,
+        iconClass: `${baseIconClass} w-[64px] h-[64px] tablet:w-[72px] tablet:h-[72px] desktop:w-[64px] desktop:h-[64px]`,
         offsets: {
           mobile: 24,
           tablet: 30,
@@ -44,8 +32,8 @@ export const getIconClasses = (type: IconClassType): IconClassSet => {
 
     default:
       return {
-        circleClass: '',
-        iconClass: '',
+        circleClass: baseCircleClass,
+        iconClass: baseIconClass,
         offsets: {
           mobile: 0,
           tablet: 0,
